+++ inst/d3js/modelStudio.js
@@ -111,11 +111,16 @@ var CLICKED_VARIABLE_NAME = variableNames[0],
     IS_BUTTON_CLICKED = false;
 
 /// set dimensions
-var margin = {top: options.margin_top, right: options.margin_right,
-              bottom: options.margin_bottom, left: options.margin_left,
-              inner: options.margin_inner,
-              small: options.margin_small, big: options.margin_big,
-              ytitle: options.margin_ytitle};
+var margin = {
+  top: options.margin_top,
+  right: options.margin_right,
+  bottom: options.margin_bottom,
+  left: options.margin_left,
+  inner: options.margin_inner,
+  small: options.margin_small,
+  big: options.margin_big,
+  ytitle: options.margin_ytitle
+};
 
 var w = options.w, h = options.h;
 
@@ -490,7 +495,7 @@ function initializeStudio() {
 
   if (facetData.length >= 1) {
     svg.selectAll('.enterChoiceButton').filter('#enterChoiceButton0').dispatch('click');
-    svg.select("#chosePlotButton0").select("#BD").dispatch('click');
+    svg.select("#chosePlotButton0").select("#FI").dispatch('click');
   }
 
   if (facetData.length >= 2) {